+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Program can download and load some common missing libraries to run games
 * Remove multiple markers (#596)
 * Add more lua functions (savestates, ff)
+* Implement ppoll
 
 ### Changed